Allow `react-scripts build --dev`
Hello,

I'll be honest, I've been going straight for the simplest solution, and I haven't actually tested it. Definitely review this before merging. I'm mostly doing this PR to try and prompt the maintainers to let such a thing exist.

I feel like there's too big a leap from `start` to `build`, as one also spawns a server (not required for my usecase), whereas the other only does production-optimized builds.

This might not be the cleanest way to do it, but I think this would allow a middle ground where one can run fast builds without spawning a server (useful when an other server is already serving the build folder).